Refactor tools and memory into dedicated modules
